% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyna.R
\name{applyna}
\alias{applyna}
\title{Apply is.na to elements in margin of array or matrix}
\usage{
applyna(x, mar = 2)
}
\arguments{
\item{x}{array, matrix, or data.frame}

\item{mar}{margin: 1=rows, 2=columns, etc.}
}
\description{
Apply is.na to elements in margin of array or matrix
}

