% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapply.R
\name{do.rep}
\alias{do.rep}
\title{Evaluate a function repeatedly over arbitrary arguments}
\usage{
do.rep(f, ..., .reps, .verbose = 1, .rep.cores = 1, .eval = T)
}
\arguments{
\item{f}{function to be evaluated}

\item{...}{Arguments passed to f}

\item{.reps}{the number of times the function should be evaluated}

\item{.verbose}{If \code{1} (default), prints a \code{.} with every completed condition. 
If \code{2}, prints the arguments corresponding to the completed condition. 
If \code{3}, prints the arguments and results of the completed condition.}

\item{.rep.cores}{Apply repeates in parallel using mclapply}

\item{.eval}{If \code{TRUE} (default), evaluates \code{f}. If \code{FALSE}, does not evaluate \code{f}.}
}
\description{
This idiom is really useful to carry out simulations, which are essentially
repeated evaluations of a function over a grid of parameter values.
}

