% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_gapply.R
\name{estimate.time}
\alias{estimate.time}
\title{Estimate time for a given number of reps}
\usage{
estimate.time(object, nreps = NULL)
}
\arguments{
\item{object}{gapply object}

\item{nreps}{number of reps to scale to}
}
\description{
Estimate time for a given number of reps
}

