% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhead.R
\name{hh}
\alias{hh}
\title{Return a small number of rows and columns}
\usage{
hh(x, nrows = 10, ncols = nrows)
}
\arguments{
\item{x}{list, matrix, or data.frame}

\item{nrow}{number of rows to return}

\item{ncol}{number of columns to return (default: nrow))}
}
\description{
Return a small number of rows and columns
}

