% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hstr.R
\name{hstr}
\alias{hstr}
\title{Compute 'str' on the first n elements or columns}
\usage{
hstr(x, n = 10, ...)
}
\arguments{
\item{x}{list, matrix or data.frame, otherwise converts to data.frame}

\item{n}{number of list elements or columns}

\item{ids}{vector of list element ids or column ids (default 1:n)}
}
\description{
Compute 'str' on the first n elements or columns
}

