% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headtail.R
\name{ht}
\alias{ht}
\title{Return a small non-overlapping number of rows and columns at beginning and end of data frame}
\usage{
ht(x, nrows = 10, ncols = nrows)
}
\arguments{
\item{x}{list, matrix or data frame}

\item{nrow}{number of rows to return (or list elements)}

\item{ncol}{number of columns to return (default: nrow))}
}
\description{
Return a small non-overlapping number of rows and columns at beginning and end of data frame
}

