% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isspecial.R
\name{is.special}
\alias{is.special}
\title{checks if \code{x} is \code{NA}, \code{NaN}, or \code{Inf}.}
\usage{
is.special(x)
}
\arguments{
\item{x}{value}
}
\description{
Returns \code{TRUE} if \code{x} is \code{NA}, \code{NaN}, or \code{Inf}.
}

