% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{list objects and sizes}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{environment (usually the global one)}

\item{n}{number of items to report}
}
\description{
Easy way to manage memory, commonly suggested to put this in .Rprofile.
Here it is in a package.
}
\references{
\url{http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
Also available in package \code{dmisc}
}

