% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing.R
\name{plotMissing}
\alias{plotMissing}
\title{plots missing values}
\usage{
plotMissing(d, main = "", varnames = NULL, minss = 50,
  transpose = FALSE, ...)
}
\arguments{
\item{d}{data}

\item{main}{title}

\item{varnames}{variable names}

\item{minss}{minimum sample size of pattern to plot}

\item{topvars}{top number of variables}
}
\description{
plots missing values
}

