% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reup.R
\name{push}
\alias{push}
\title{Push changes to tracked files to master branch on github}
\usage{
push(msg, f = NULL)
}
\arguments{
\item{m}{commit message}
}
\description{
Push changes to tracked files to master branch on github
}

