% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putArgs.R
\name{put.args}
\alias{put.args}
\title{Assign default args of a function into global environment}
\usage{
put.args(f, env = .GlobalEnv)
}
\arguments{
\item{f}{function name}

\item{env}{environment, default is Global}
}
\value{
Called for the side effect
}
\description{
useful for interactive debugging of functions, call this then step through the function.
}

