% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwtables.R
\name{pw.tables}
\alias{pw.tables}
\title{Compute all pairwise tables for data}
\usage{
pw.tables(x)
}
\arguments{
\item{x}{matrix or data.frame}
}
\value{
list of all pairwise tables
}
\description{
Compute all pairwise tables for data
}

