% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_setup.R
\name{setup}
\alias{setup}
\title{setup sge simulation}
\usage{
setup(object, dir = "", script.name = "doone.R", ncores = 1, nreps = 1)
}
\arguments{
\item{dir}{directory name relative to the current working directory, ends in '/'}
}
\description{
setup sge simulation
}

