% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shead.R
\name{sh}
\alias{sh}
\title{Sample rows and columns of a matrix}
\usage{
sh(x, nrow = 10, ncol = nrow, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{matrix or data frame, otherwise converts to a matrix}

\item{nrow}{number of rows to sample}

\item{ncol}{number of columns to sample (defaults to nrow)}

\item{rows}{vector of row ids. If specified, rows are selected by row id and not sampled}

\item{cols}{vector of column ids. If specified, columns are selected by column id and not sampled}
}
\description{
Sample rows and columns of a matrix
}

