% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sstr.R
\name{sstr}
\alias{sstr}
\title{Print the structure of a sample of columns or list elements}
\usage{
sstr(x, n = 10, ...)
}
\arguments{
\item{x}{list, matrix or data frame, otherwise converts to data.frame}

\item{n}{number of columns to sample}
}
\description{
Print the structure of a sample of columns or list elements
}

