% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapply.R
\name{summary.gapply}
\alias{summary.gapply}
\title{\code{summary} method for class \code{"gapply"}}
\usage{
\method{summary}{gapply}(object, nreps = NULL)
}
\arguments{
\item{object}{gapply result object}

\item{nreps}{number of reps to scale to}
}
\value{
Prints the means over all reps for each condition and returns it invisibly. Also prints the estimated time to scale up to x reps
}
\description{
\code{summary} method for class \code{"gapply"}
}

