% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le.R
\name{ule}
\alias{ule}
\title{Return the ith element of a nested list, and unlist}
\usage{
ule(x, i = 1)
}
\arguments{
\item{x}{(nested) list}

\item{i}{index of the element to return}
}
\description{
Return the ith element of a nested list, and unlist
}

