% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2symmetric.R
\name{vec2sym}
\alias{vec2sym}
\title{Convert vector to symmetric matrix}
\usage{
vec2sym(x, diagonal = NULL, lower = TRUE, byrow = FALSE)
}
\arguments{
\item{x}{vector containing upper or lower triangular elements of a symmetric matrix}

\item{lower}{x is from the lower triangle (default = \code{TRUE})}

\item{byrow}{the elements in x are ordered row-wise (default = \code{FALSE})}

\item{If}{NULL, x contains diagonal elements. Otherwise a value or vector of the appropriate length to be placed on the diagonal.}
}
\value{
Symmetric matrix
}
\description{
Reverse operation of \code{upper.tri} or \code{lower.tri}. Given a vector of 
upper or lower triangular elements of a matrix 
(optionally including the diagonal elements), returns the
corresponding symmetric matrix. The elements of the vector can be arranged by
column (the default) or by row.
}
\details{
Note that if x is a vector of the lower triagular elements given by column, 
this is equivalent to the upper triangular elements given by row. Similarly, if 
x is a vector of the lower triangular elements given by row, this is equivalent
to the upper triangular elements given by column.
}
\examples{
x <- c(1,1,1,2,2,3)
check <- matrix(c(1,1,1,1,1,1,2,2,1,2,1,3,1,2,3,1),4,4) 
identical(vec2sym(x,diag=1,lower=T,byrow=F), check)
x <- c(1,1,1,2,2,3)
check <- matrix(c(1,1,1,1,2,2,1,2,3),3,3)
identical(vec2sym(x,lower=T,byrow=F),check)
}

