% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_setup.R
\name{setup}
\alias{setup}
\title{setup sge simulation}
\usage{
setup(object, dir = "", .reps = 1, .chunks = 1, .mc.cores = 1,
  .verbose = 1, .script.name = "doone.R")
}
\arguments{
\item{dir}{directory name relative to the current working directory, ends in '/'}
}
\description{
setup sge simulation
}

