% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2symmetric.R
\name{vec2sym}
\alias{vec2sym}
\title{convert vector to symmetric matrix}
\usage{
vec2sym(x, diagonal = NULL, lower = TRUE, byrow = FALSE)
}
\arguments{
\item{x}{vector containing upper or lower triangular elements of a symmetric matrix}

\item{diagonal}{value or vector of elements to place on the diagonal}

\item{lower}{x is from the lower triangle (default = \code{TRUE})}

\item{byrow}{the elements in x are ordered row-wise (default = \code{FALSE})}
}
\description{
Given a vector of upper or lower triangular elements of a matrix, returns the
corresponding symmetric matrix. The elements of the vector can be arranged by
column (the default) or by row.
}

